fix of the array type cast bugs

+++ test/test/arrayTest2.vb
@@ -8,7 +8,7 @@
 
     Declare Function print Lib "console" Alias "log" (x As Double)
 
-    Function returnArrayTest() As Integer()
+    Function returnArrayTest() As Single()
         Dim x As Double = data(1)
 
         Return {x, 0, 35, 78345, 34, 534, 53}
